WIP - removed 'latestOnly' from LinearStateQueryCriteria

+++ core/src/main/kotlin/net/corda/core/node/services/Services.kt
@@ -117,7 +117,7 @@ class Vault<out T : ContractState>(val states: Iterable<StateAndRef<T>>) {
     data class Page<out T : ContractState>(val states: List<StateAndRef<T>>,
                                            val statesMetadata: List<Vault.StateMetadata>,
                                            val pageable: PageSpecification,
-                                           val totalStatesAvailable: Long)
+                                           val totalStatesAvailable: Int)
 
     @CordaSerializable
     data class StateMetadata(val ref: StateRef,
